Backport our work around for bad data in some column families.
We run into OOM errors on some column families when bad rows get
written.  This is not a great work around as it just dumps the rows and
assumes that data that is too large is bad.

A more full solution will be to handle rows on a column by column basis,
although in some cases rows can still be longer than our final
serialized string due to limitations on memory.